Better options to disable auto detection of mobile devices
New attribute "no_auto_mobile" that can be added to the shortcode. Will add
a new option in the form for the user to select whether he wants to join
using the mobile client or not. Will disable the auto detection of mobile
devices and use the client selected by the user.