Change the option to disable auto detection of mobile devices
The attribute is now called "show_mobile_options". If set, the option to
select if the mobile client should be used or not will be shown and the
auto detection will be disabled. The checkbox is automatically checked if
the user is in a mobile device, but the user can unchecked it and force the
use of the desktop client. The other way around (forcing the mobile client)
also works.